feat: solhint configuration solidity lint

+++ nightfall-deployer/contracts/Config.sol
@@ -3,9 +3,9 @@
 pragma solidity ^0.8.0;
 
 contract Config {
-  uint constant REGISTRATION_BOND = 10 wei; // TODO owner can update
-  uint constant BLOCK_STAKE = 1 wei;
-  uint constant ROTATE_PROPOSER_BLOCKS = 4;
-  uint constant COOLING_OFF_PERIOD = 1 weeks;
-  bytes32 constant ZERO = bytes32(0);
+    uint256 constant REGISTRATION_BOND = 10 wei; // TODO owner can update
+    uint256 constant BLOCK_STAKE = 1 wei;
+    uint256 constant ROTATE_PROPOSER_BLOCKS = 4;
+    uint256 constant COOLING_OFF_PERIOD = 1 weeks;
+    bytes32 constant ZERO = bytes32(0);
 }
